requirements.txt
appdirs>=1.4.4
attrs>=23.1.0
black>=23.1.0
click>=8.1.3
cycler>=0.11.0
kiwisolver>=1.3.1
matplotlib>=3.8.0
nose>=1.3.7
numpy>=1.24.0
opencv-contrib-python>=4.8.0.76
pandas>=1.5.3
Pillow>=9.5.0
pyparsing>=3.1.0
python-dateutil>=2.8.2
pytz>=2023.3
six>=1.16.0
toml>=0.10.2
virtualenv>=20.21.0
xlrd>=2.0.1
xmltodict>=0.13.0
yagmail>=0.15.293
yapf>=0.40.1
